build bundled libraries dynamically
This avoid the need to go through a configure
file to copy the right version of the standard
library.

Cross-version support: using codept build with OCaml X
to compute dependencies of OCaml Y with the right stdlib
modules would require to functorize the bundle module.

The module is currently handwritten but is much easier
to read and hopefully maintain that the various
generated files.

A validator might be useful.