Introduce the new Modname module
The Modname module is like the Name but with an abstract type. The idea
behind is to provide a nice API and codept as a library. The abstraction
lets us to keep some invariants described into the documentation.

The goal is to use this abstract type for any module names instead of
the Name.t which can refers to several things (path, namespace, module
path, etc.)