[test] Update how we compare results and what we expect into tests/run.ml
The codept solver returns module names instead of a basic strings. To
compare results and what we abstract, we must cast results to strings
(due to the fact that Modname.t is an abstract type now). The comparison
is also updated due to the new definition of Namespaced.t. We must use
Namespaced.compare instead of Stdlib.(=) to select only one view of a
namespace (the module path view instead of the file path view).