Introduce the new Unitname module
The objective of Unitname is to keep the path of a given module. Due to
the namespaced view, it's possible that a path has a signification for
the solver (as a part of the module path). To ensure the bijection on
some elements, we must keep the file path for any modules which are
concretized by a file. Unitname offers two views:
- the module name (introduced by the Modname module)
- the file path used to construct the Unitname

By this way, we ensure the bijection on Unitname elements and we provide
these views depending on the context.